+++ app-customer/src/main/java/com/petproject/appcustomer/domain/ports/in/UpdateCustomerUseCase.java
@@ -3,5 +3,5 @@ package com.petproject.appcustomer.domain.ports.in;
 import com.petproject.appcustomer.domain.models.in.CustomerEntity;
 
 public interface UpdateCustomerUseCase {
-    CustomerEntity updateCustomer(Long id, CustomerEntity customerEntity);
+    CustomerEntity updateCustomer(Integer id, CustomerEntity customerEntity);
 }
